requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
python-telegram-bot
python-dotenv